Refactor Codey to be more modular and secure.
Extracts code generation into the RandomCodeGenerator class
to clean up the Secret model.

Moves random code generation from Secret to Verification. The
Secret now clears the @code and @data after its encrypted for
additional security and reduces the risk of a developer accidentally
keeping secrets around after its encrypted.

The code is persisted after a Secret is encrypted in the Authorization
object. This is so the developer can still read the code to send to
the end user.